Allow specs to specify additional cool down times

+++ src/main/java/com/skelril/aurora/items/specialattack/SpecialAttack.java
@@ -42,7 +42,7 @@ public abstract class SpecialAttack {
         }
     }
 
-    public int coolDown() {
+    public long coolDown() {
 
         return 0;
     }
